requirements.txt
streamlit
openai
langchain
chromadb    
langchain-community
faiss-cpu
python-dotenv
pypdf
pypdf2
streamlit-chat
tiktoken
python-docx
numpy

